Unify TrackerWindow implementations
